Device Info: Read new property to indicate custom build version [3/3]
The 'official' Security string has always been taken from the LineageOS
sources, as Android security patches are frequently back-ported by the
members of the community.

Recently, certain ASB patches could not (yet) be backported, whilst it is
acknowledged, that they are relevant. Hence the official stance of the
LineageOS directors is to continue backporting any ASB patches, but not
to lift the security string in order to be transparent about the fact,
that the formal criteria to declare a patch level haven't been met.
The official patch level will be lifted accordingly, once the missing
patches will have been successfully backported and merged.

To consider that (not lifting the patch), but at the same time indicate,
that despite the missing patches, further backporting of security issues
is ongoing, a new "Custom build release" property 'ro.cm.custom_version'
is introduced to show this next to the "official" security string.

This commit provides a new preference to show this additional information.

Change-Id: I14f41ab305acaeaaab42596046498696d8e6bbda